// src/types/enums.ts
export enum RecallStatus {
  RECALL_INITIATED = "RECALL_INITIATED",
  NOTIFICATIONS_SENT = "NOTIFICATIONS_SENT",
  RETURNS_IN_PROGRESS = "RETURNS_IN_PROGRESS",
  RECALL_CLOSED = "RECALL_CLOSED"
}

export enum ObligationStatus {
  PENDING = "PENDING",
  ACKNOWLEDGED = "ACKNOWLEDGED",
  RETURNED = "RETURNED"
}